requirements.txt
fastapi
uvicorn
psycopg2-binary
python-decouple
passlib[bcrypt]
PyJWT
